# DRY out quantum_mallows function