resolves #49

I added a helper function `sample_geometric_2` to sample a random number with respect to the desired distribution.

I removed the three versions of `quantum_mallows_int`, `quantum_mallows_float`, and `quantum_mallows_bigint` since they were almost identical (except for the random number generation for which I used the `sample_geometric_2` function).