@Krastanov 2023-04-01
Thank you! Benchmark shows this is a bit slower for very small tableaux and a bit faster for large tableaux, but overall it is certainly cleaner. Will merge shortly (with a small type-stability improvement).

After:

```
julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 10)
BenchmarkTools.Trial: 10000 samples with 196 evaluations.
 Range (min … max):  466.898 ns …  14.471 μs  ┊ GC (min … max): 0.00% … 94.38%
 Time  (median):     558.398 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   598.363 ns ± 471.082 ns  ┊ GC (mean ± σ):  3.05% ±  3.74%
 Memory estimate: 384 bytes, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 100)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  10.941 μs … 33.333 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     11.782 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.525 μs ±  1.573 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.86 KiB, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 600)
BenchmarkTools.Trial: 9838 samples with 1 evaluation.
 Range (min … max):  212.750 μs … 64.049 ms  ┊ GC (min … max):  0.00% … 67.87%
 Time  (median):     257.495 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   507.196 μs ±  3.771 ms  ┊ GC (mean ± σ):  31.28% ±  4.17%
 Memory estimate: 186.20 KiB, allocs estimate: 3842.
```

Before: 

```
julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 10)
BenchmarkTools.Trial: 10000 samples with 200 evaluations.
 Range (min … max):  397.695 ns …  11.452 μs  ┊ GC (min … max): 0.00% … 94.07%
 Time  (median):     475.995 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   503.176 ns ± 402.976 ns  ┊ GC (mean ± σ):  3.12% ±  3.74%
 Memory estimate: 384 bytes, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 10)
BenchmarkTools.Trial: 10000 samples with 200 evaluations.
 Range (min … max):  397.195 ns …  12.656 μs  ┊ GC (min … max): 0.00% … 94.39%
 Time  (median):     474.645 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   504.529 ns ± 442.512 ns  ┊ GC (mean ± σ):  3.41% ±  3.75%
 Memory estimate: 384 bytes, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 10)
BenchmarkTools.Trial: 10000 samples with 200 evaluations.
 Range (min … max):  398.700 ns …  12.046 μs  ┊ GC (min … max): 0.00% … 92.02%
 Time  (median):     477.450 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   517.305 ns ± 430.911 ns  ┊ GC (mean ± σ):  3.22% ±  3.74%
 Memory estimate: 384 bytes, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 100)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   8.977 μs … 25.478 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     11.131 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   11.811 μs ±  1.621 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.86 KiB, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 100)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   8.997 μs … 32.090 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     11.241 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   11.870 μs ±  1.535 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.86 KiB, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 100)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   9.087 μs … 25.328 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     12.063 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   12.229 μs ±  1.833 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 1.86 KiB, allocs estimate: 4.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 600)
BenchmarkTools.Trial: 6298 samples with 1 evaluation.
 Range (min … max):  298.833 μs … 81.968 ms  ┊ GC (min … max):  0.00% … 64.37%
 Time  (median):     361.601 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   790.675 μs ±  5.376 ms  ┊ GC (mean ± σ):  34.94% ±  5.09%
 Memory estimate: 256.66 KiB, allocs estimate: 6570.

julia> @benchmark QuantumClifford.quantum_mallows(Random.GLOBAL_RNG, 600)
BenchmarkTools.Trial: 6446 samples with 1 evaluation.
 Range (min … max):  300.195 μs … 75.819 ms  ┊ GC (min … max):  0.00% … 64.52%
 Time  (median):     361.270 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   776.407 μs ±  5.332 ms  ┊ GC (mean ± σ):  35.12% ±  5.06%
 Memory estimate: 256.66 KiB, allocs estimate: 6570.
```